requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
distlib==0.3.4
filelock==3.7.1
Flask==2.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
platformdirs==2.5.2
PyYAML==6.0
six==1.16.0
urllib3==1.26.9
virtualenv==20.14.1
waitress==2.1.2
Werkzeug==2.1.2
zipp==3.8.0
